refactor: (#14) kotlin reactor extension으롤 가독성 개선
