+++ src/main/kotlin/org/example/chitchatserver/ChitchatServerApplication.kt
@@ -3,9 +3,9 @@ package org.example.chitchatserver
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
-import org.springframework.data.mongodb.config.EnableReactiveMongoAuditing
 
-@EnableReactiveMongoAuditing
+internal const val BASE_PACKAGE = "org.example.chitchatserver"
+
 @ConfigurationPropertiesScan
 @SpringBootApplication
 class ChitchatServerApplication
