+++ src/main/kotlin/org/example/chitchatserver/domain/auth/persistence/RefreshTokenRepository.kt
@@ -1,0 +1,7 @@
+package org.example.chitchatserver.domain.auth.persistence
+
+import reactor.core.publisher.Mono
+
+interface RefreshTokenRepository {
+    fun save(refreshTokenEntity: RefreshTokenEntity): Mono<Boolean>
+}
