--- src/main/kotlin/org/example/chitchatserver/domain/auth/persistence/repository/RefreshTokenRepository.kt
@@ -1,8 +1,0 @@
-package org.example.chitchatserver.domain.auth.persistence.repository
-
-import org.example.chitchatserver.domain.auth.persistence.RefreshTokenEntity
-import reactor.core.publisher.Mono
-
-interface RefreshTokenRepository {
-    fun save(refreshTokenEntity: RefreshTokenEntity): Mono<Boolean>
-}
